Reduce budget_program and config_program into boilerplate crates
