+++ programs/budget_program/Cargo.toml
@@ -9,17 +9,10 @@ homepage = "https://solana.com/"
 edition = "2018"
 
 [dependencies]
-bincode = "1.1.2"
-chrono = { version = "0.4.0", features = ["serde"] }
 log = "0.4.2"
-serde = "1.0.89"
-serde_derive = "1.0.89"
 solana-budget-api = { path = "../budget_api", version = "0.13.0" }
 solana-logger = { path = "../../logger", version = "0.13.0" }
 solana-sdk = { path = "../../sdk", version = "0.13.0" }
-
-[dev-dependencies]
-solana-runtime = { path = "../../runtime", version = "0.13.0" }
 
 [lib]
 name = "solana_budget_program"
