+++ programs/budget_program/src/lib.rs
@@ -1,5 +1,3 @@
-mod budget_processor;
-
-use crate::budget_processor::process_instruction;
+use solana_budget_api::budget_processor::process_instruction;
 
 solana_sdk::process_instruction_entrypoint!(process_instruction);
