+++ programs/config_api/Cargo.toml
@@ -13,7 +13,11 @@ bincode = "1.1.2"
 log = "0.4.2"
 serde = "1.0.89"
 serde_derive = "1.0.89"
+solana-logger = { path = "../../logger", version = "0.13.0" }
 solana-sdk = { path = "../../sdk", version = "0.13.0" }
+
+[dev-dependencies]
+solana-runtime = { path = "../../runtime", version = "0.13.0" }
 
 [lib]
 name = "solana_config_api"
