+++ programs/config_api/src/lib.rs
@@ -2,6 +2,7 @@ use serde::Serialize;
 use solana_sdk::pubkey::Pubkey;
 
 mod config_instruction;
+pub mod config_processor;
 
 pub use config_instruction::ConfigInstruction;
