+++ programs/config_program/Cargo.toml
@@ -10,17 +10,9 @@ edition = "2018"
 
 [dependencies]
 log = "0.4.2"
-solana-logger = { path = "../../logger", version = "0.13.0" }
-solana-metrics = { path = "../../metrics", version = "0.13.0" }
-solana-sdk = { path = "../../sdk", version = "0.13.0" }
 solana-config-api = { path = "../config_api", version = "0.13.0" }
-
-[dev-dependencies]
-bincode = "1.1.2"
-serde = "1.0.89"
-serde_derive = "1.0.89"
 solana-logger = { path = "../../logger", version = "0.13.0" }
-solana-runtime = { path = "../../runtime", version = "0.13.0" }
+solana-sdk = { path = "../../sdk", version = "0.13.0" }
 
 [lib]
 name = "solana_config_program"
